[FEAT] Post 기능 완성

+++ app/routers/board.py
@@ -4,6 +4,7 @@
 # @author bnbong bbbong9@gmail.com
 # --------------------------------------------------------------------------
 from logging import getLogger
+from typing import Tuple
 
 from fastapi import APIRouter, Depends, Query
 from sqlalchemy import select
@@ -16,6 +17,8 @@ from app.helper.exceptions import InternalException, ErrorCode
 from app.schemas import board as schemas
 from app.crud import board as crud
 from app.utils import constant
+from ._check import check_user, check_user_is_admin
+
 
 log = getLogger(__name__)
 board_router = APIRouter(prefix="/board")
@@ -58,11 +61,16 @@ async def read_board(board_id: int, db: AsyncSession = Depends(database.get_db))
     response_model=schemas.BoardSchema,
     status_code=201,
     summary="게시판 생성",
-    description="새로운 게시판을 생성합니다.",
+    description="새로운 게시판을 생성합니다.\n\nAdmin(관리자) 이상의 권한이 필요합니다.",
 )
 async def create_board(
-    board: schemas.BoardCreateSchema, db: AsyncSession = Depends(database.get_db)
+    board: schemas.BoardCreateSchema,
+    db: AsyncSession = Depends(database.get_db),
+    request_user: Tuple[str, int] = Depends(check_user),
 ):
+    user_pk, role_pk = request_user
+    await check_user_is_admin(db=db, role_pk=role_pk)
+
     db_board = await crud.create_board(db, board)
     return db_board
 
@@ -71,19 +79,23 @@ async def create_board(
     "/{board_id}",
     response_model=schemas.BoardSchema,
     summary="게시판 수정 (전체 업데이트)",
-    description="지정한 게시판의 데이터를 전체적으로 수정합니다. (부분 업데이트도 지원합니다)",
+    description="지정한 게시판의 데이터를 전체적으로 수정합니다. (부분 업데이트도 지원합니다)\n\nAdmin(관리자) 이상의 권한이 필요합니다.",
 )
 @board_router.patch(
     "/{board_id}",
     response_model=schemas.BoardSchema,
     summary="게시판 수정 (부분 업데이트)",
-    description="지정한 게시판의 데이터를 부분적으로 수정합니다.",
+    description="지정한 게시판의 데이터를 부분적으로 수정합니다.\n\nAdmin(관리자) 이상의 권한이 필요합니다.",
 )
 async def update_board(
     board_id: int,
     board: schemas.BoardUpdateSchema,
     db: AsyncSession = Depends(database.get_db),
+    request_user: Tuple[str, int] = Depends(check_user),
 ):
+    user_pk, role_pk = request_user
+    await check_user_is_admin(db=db, role_pk=role_pk)
+
     db_board = await crud.update_board(db, board_id, board)
     if db_board is None:
         raise InternalException("해당 게시판을 찾을 수 없습니다.", ErrorCode.NOT_FOUND)
@@ -94,9 +106,16 @@ async def update_board(
     "/{board_id}",
     status_code=204,
     summary="게시판 삭제",
-    description="지정한 게시판을 삭제합니다.",
+    description="지정한 게시판을 삭제합니다.\n\nAdmin(관리자) 이상의 권한이 필요합니다.",
 )
-async def delete_board(board_id: int, db: AsyncSession = Depends(database.get_db)):
+async def delete_board(
+    board_id: int,
+    db: AsyncSession = Depends(database.get_db),
+    request_user: Tuple[str, int] = Depends(check_user),
+):
+    user_pk, role_pk = request_user
+    await check_user_is_admin(db=db, role_pk=role_pk)
+
     deleted_id = await crud.delete_board(db, board_id)
     if deleted_id is None:
         raise InternalException("해당 게시판을 찾을 수 없습니다.", ErrorCode.NOT_FOUND)
